Split NoMethodsInPropertySetup
