+++ Source/Moq.Analyzers.Test/NoMethodsInPropertySetupAnalyzerTests.ShouldFailWhenMethodsUsePropertySetup.verified.txt
@@ -1,6 +1,6 @@
 ﻿Diagnostic 1
 	Id: Moq1101
-	Location: SourceFile(Test0.cs[307..317))
+	Location: SourceFile(Test0.cs[311..321))
 	Highlight: x.Method()
 	Lines: mock.SetupGet(x => x.Method());
 	Severity: Warning
@@ -8,7 +8,7 @@
 
 Diagnostic 2
 	Id: Moq1101
-	Location: SourceFile(Test0.cs[347..357))
+	Location: SourceFile(Test0.cs[351..361))
 	Highlight: x.Method()
 	Lines: mock.SetupSet(x => x.Method());
 	Severity: Warning
