+++ Source/Moq.Analyzers.Test/NoMethodsInPropertySetupAnalyzerTests.ShouldPassWhenPropertiesUsePropertySetup.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
