+++ Source/Moq.Analyzers.Test/NoMethodsInPropertySetupAnalyzerTests.cs
@@ -9,10 +9,17 @@ namespace Moq.Analyzers.Test;
 public class NoMethodsInPropertySetupAnalyzerTests : DiagnosticVerifier
 {
     [Fact]
-    public Task Test()
+    public Task ShouldPassWhenPropertiesUsePropertySetup()
     {
-        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/NoMethodsInPropertySetup.cs")));
+        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/NoMethodsInPropertySetup.Good.cs")));
     }
+
+    [Fact]
+    public Task ShouldFailWhenMethodsUsePropertySetup()
+    {
+        return Verify(VerifyCSharpDiagnostic(File.ReadAllText("Data/NoMethodsInPropertySetup.Bad.cs")));
+    }
+
 
     protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
     {
